FX-T-28: - builder for format

+++ src/RxBim.Tools.TableBuilder/Models/Styles/CellTextFormatStyle.cs
@@ -25,7 +25,7 @@
         /// <summary>
         /// The font size of the text.
         /// </summary>
-        public int? TextSize { get; set; }
+        public double? TextSize { get; set; }
 
         /// <summary>
         /// Use word wrap for long text.
